Consolidate backend ci-conductor reporting into release/ci-conductor
Stand up a shared, bun-run TS package for reporting test failures to
ci-conductor, built around an adapter pattern: each test suite transforms
its source data into one canonical shape and the shared core posts it. This
round ships the core plus the backend (JUnit) adapter, superseding the
standalone zero-dep CommonJS reporter.

- release/ci-conductor/: contract (canonical shape, types only), identity
  (run context), transport (POST /webhooks/failed-tests), and the junit
  adapter (hawk XML -> canonical), with the entrypoint run directly via bun.
- Carry over the quote-aware JUnit attribute parsing so Clojure test names
  with unescaped angle brackets aren't silently dropped, plus its regression
  test (now a bun test).
- Rewire backend.yml and the test-driver composite to set up bun and run the
  TS reporter; drop the old bin/ci-conductor reporter and root npm script.
- Add a release-tests.yml job to type-check and run the package's tests.
