+++ .github/actions/test-driver/action.yml
@@ -48,6 +48,14 @@ runs:
       # Even if this step fails we let it pass, and then `trunk-io/analytics-uploader` determines
       # whether to fail or to pass the whole job, based on the quarantine list.
       continue-on-error: true
+    - name: Install bun (for ci-conductor reporting)
+      # Only when we're about to report — keeps clean runs fast. The driver jobs
+      # are JDK-only, so bun isn't otherwise present.
+      if: ${{ always() && steps.run-test.outcome != 'success' }}
+      uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
+      with:
+        bun-version: "1.3.14"
+        token: ${{ github.token }}
     - name: Report failed tests to ci-conductor
       # A visibility step: parse the raw JUnit hawk just wrote and POST failures to
       # ci-conductor. Positioned right after the test run (before the report/upload
@@ -55,7 +63,9 @@ runs:
       # the report below (run on anything that isn't a clean pass), and
       # continue-on-error so it can never change the job outcome — Trunk Quarantine
       # remains the sole arbiter. CI_CONDUCTOR_BASE_URL / _WEBHOOK_SECRET and the
-      # resolved JOB_ID arrive via the workflow/job env. See DEV-2224.
+      # resolved JOB_ID arrive via the workflow/job env. The reporter is
+      # zero-runtime-dep (bun built-ins only), so it just runs the TS directly —
+      # no `bun install`. See DEV-2224 / DEV-2235.
       if: ${{ always() && steps.run-test.outcome != 'success' }}
       shell: bash
       continue-on-error: true
@@ -68,7 +78,7 @@ runs:
         # so ci-conductor buckets line up with it. Only workflows that also set
         # CI_CONDUCTOR_BASE_URL actually POST; others no-op.
         CI_CONDUCTOR_TEST_SUITE: ${{ inputs.junit-name }}
-      run: node bin/ci-conductor/report-junit-failures.js
+      run: bun release/ci-conductor/src/report-junit.ts
     - name: Publish Test Report (JUnit)
       uses: dorny/test-reporter@a43b3a5f7366b97d083190328d2c652e1a8b6aa2 # v3.0.0
       # As per the documentation: https://docs.github.com/en/actions/reference/workflows-and-actions/contexts#steps-context
