+++ .github/workflows/backend.yml
@@ -202,6 +202,15 @@ jobs:
         # whether to fail or to pass the whole job, based on the quarantine list.
         continue-on-error: true
 
+      - name: Install bun (for ci-conductor reporting)
+        # Only when we're about to report — keeps clean runs fast. The
+        # backend/driver jobs are JDK-only, so bun isn't otherwise present.
+        if: ${{ always() && steps.run-java-tests.outcome != 'success' }}
+        uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
+        with:
+          bun-version: "1.3.14"
+          token: ${{ github.token }}
+
       - name: Report failed tests to ci-conductor
         # A visibility step: parse the raw JUnit hawk just wrote and POST failures
         # to ci-conductor. Positioned between the test run and "Upload Test Results"
@@ -209,7 +218,8 @@ jobs:
         # log is easy to find. continue-on-error so it can never change the job
         # outcome — Trunk Quarantine remains the sole arbiter. CI_CONDUCTOR_BASE_URL
         # / _WEBHOOK_SECRET and the resolved JOB_ID arrive via the workflow/job env.
-        # See DEV-2224.
+        # The reporter is zero-runtime-dep (bun built-ins only), so it just runs
+        # the TS directly — no `bun install`. See DEV-2224 / DEV-2235.
         if: ${{ always() && steps.run-java-tests.outcome != 'success' }}
         continue-on-error: true
         env:
@@ -217,7 +227,7 @@ jobs:
           COMMIT_SHA: ${{ github.event.pull_request.head.sha || github.sha }}
           TARGET_BRANCH: ${{ github.event.pull_request.base.ref || github.ref_name }}
           CI_CONDUCTOR_TEST_SUITE: be-tests-java-${{ matrix.java-version }}-${{ matrix.job.test-group-name }}
-        run: node bin/ci-conductor/report-junit-failures.js
+        run: bun release/ci-conductor/src/report-junit.ts
 
       - name: Upload Test Results
         uses: ./.github/actions/upload-test-results
