+++ release/ci-conductor/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "ci-conductor",
+  "version": "1.0.0",
+  "description": "Shared CI Conductor test-failure reporting. Adapter pattern: each test suite (be/fe/e2e) transforms its source data into one canonical shape, and the shared core posts it to ci-conductor. This round ships the backend (JUnit) adapter + the /webhooks/failed-tests transport.",
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "report:junit": "bun src/report-junit.ts",
+    "test": "bun test",
+    "type-check": "tsc --noEmit"
+  },
+  "author": "Metabase",
+  "license": "AGPL-3.0-or-later",
+  "devDependencies": {
+    "@types/bun": "^1.3.14",
+    "@types/node": "^22.10.0",
+    "typescript": "^5.9.3"
+  }
+}
